+++ src/RestApi/MapRestController.php
@@ -3,26 +3,32 @@
 namespace Foodsharing\RestApi;
 
 use Foodsharing\Lib\Session;
+use Foodsharing\Modules\Core\DBConstants\Region\RegionPinStatus;
 use Foodsharing\Modules\Core\DBConstants\Store\CooperationStatus;
 use Foodsharing\Modules\Core\DBConstants\Store\TeamStatus;
 use Foodsharing\Modules\Map\MapGateway;
+use Foodsharing\Modules\Region\RegionGateway;
 use FOS\RestBundle\Controller\AbstractFOSRestController;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Request\ParamFetcher;
 use OpenApi\Annotations as OA;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 
 class MapRestController extends AbstractFOSRestController
 {
 	private MapGateway $mapGateway;
+	private RegionGateway $regionGateway;
 	private Session $session;
 
 	public function __construct(
 		MapGateway $mapGateway,
+		RegionGateway $regionGateway,
 		Session $session
 	) {
 		$this->mapGateway = $mapGateway;
+		$this->regionGateway = $regionGateway;
 		$this->session = $session;
 	}
 
@@ -80,5 +86,29 @@ class MapRestController extends AbstractFOSRestController
 		}
 
 		return $this->handleView($this->view($markers, 200));
+	}
+
+	/**
+	 * Returns the data for the bubble of a community marker on the map.
+	 *
+	 * @OA\Response(response="200", description="Success.")
+	 * @OA\Response(response="404", description="The region does not exist or does not have a community description.")
+	 * @OA\Tag(name="map")
+	 *
+	 * @Rest\Get("map/regions/{regionId}")
+	 * @Rest\QueryParam(name="regionId", requirements="\d+", nullable=true, description="Region for which to return the description")
+	 */
+	public function getRegionBubbleAction(int $regionId): Response
+	{
+		$region = $this->regionGateway->getRegion($regionId);
+		$pin = $this->regionGateway->getRegionPin($regionId);
+		if (empty($pin) || $pin['status'] != RegionPinStatus::ACTIVE) {
+			throw new NotFoundHttpException('region does not exist or its pin is not active');
+		}
+
+		return $this->handleView($this->view([
+			'name' => $region['name'],
+			'description' => $pin['desc']
+		], 200));
 	}
 }